+++ cmake/nfxSiliconSources.cmake
@@ -3,6 +3,9 @@
 #==============================================================================
 
 set(NFX_SILICON_SOURCES
+    # Discrete components
+    src/discrete/PowerSupply.cpp
+
     # Signal layer
     src/signal/Bus.cpp
     src/signal/Pin.cpp