+++ include/nfx/Silicon.h
@@ -17,8 +17,12 @@
 #include "nfx/silicon/component/IInspectable.h"
 #include "nfx/silicon/component/IPassive.h"
 
+// Discrete components
+#include "nfx/silicon/discrete/PowerSupply.h"
+
 // Signal layer
 #include "nfx/silicon/signal/Bus.h"
 #include "nfx/silicon/signal/Pin.h"
+#include "nfx/silicon/signal/PowerRail.h"
 #include "nfx/silicon/signal/Signal.h"
 #include "nfx/silicon/signal/Wire.h"